+++ src/main/java/com/example/lifeonhana/repository/LoanProductRepository.java
@@ -1,7 +1,6 @@
 package com.example.lifeonhana.repository;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -11,6 +10,4 @@ import com.example.lifeonhana.entity.Product;
 @Repository
 public interface LoanProductRepository extends JpaRepository<Product, Long> {
 	List<Product> findByCategory(Product.Category category);
-
-	Optional<Product> findByProductId(Long productId);
 }
