+++ src/main/java/com/example/lifeonhana/repository/ProductRepository.java
@@ -1,5 +1,7 @@
 package com.example.lifeonhana.repository;
 
+import java.util.Optional;
+
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,5 +11,7 @@ import com.example.lifeonhana.entity.Product;
 
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Long> {
-	Page<Product> findByCategory(Product.Category productCategory, Pageable pageable) ;
+	Page<Product> findByCategory(Product.Category productCategory, Pageable pageable);
+
+	Optional<Product> findByProductId(Long productId);
 }
